Debug Menu + Overheated Bug Fixes
- Added a Command Window Menu (open/close by pressing ` ).
This is a temporary addition for Milestone 1, so the modifiers can be shown easily (as currently I don't think they can be??? We don't have the functionality of loser picking a power up yet. This is mainly added to avert the possible "It says here that the Overheated, Power-Saving and Overcharged Modifiers are completed for Milestone 1. Can you show them?" questions 💀 ). Otherwise, the alternative would be brute forcing the Modifier Menu to give a modifier to x player and that's torturous LOL. Just ease of access until things are implemented. Can and most likely will be removed later.
- Fixed a bunch of bugs with the Overheated modifier (burn debuff not applying correctly to the player, burn debuff only working once then never again, burn damage not applying due to the interval being larger than the time the debuff lasts for LOL)
- Fixed the bug with Keyboard and Mouse counting as separate controllers. Now they read as one :P